+++ lib/RGBLedManager/src/RGBLedManager.h
@@ -5,12 +5,12 @@ using namespace std;
 
 class RGBLedManager {
     public:
-    RGBLedManager(int redPin, int greenPin, int bluePin);
+    RGBLedManager(uint8_t redPin,uint8_t greenPin, uint8_t bluePin);
     void setLed(String hexColor);
 
     private:
-    int redPin;
-    int greenPin;
-    int bluePin;
+    uint8_t redPin;
+    uint8_t greenPin;
+    uint8_t bluePin;
     uint8_t red, green, blue;
 };
